assistant-api/requirements.txt
# Core
requests==2.28.2

# Django
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-filter==22.1

# Libraries
python-dotenv==0.21.1
markdown==3.4.1
python-slugify==8.0.1

# OpenAI
openai==0.27.4

# Cache
django-redis==5.2.0
